chore: Rename DialectContext
